chore(lint): three small cleanups surfaced by ruff
Mechanical fixes for the remaining ruff findings after the
configuration tuning in 3352831. None change behavior; pytest
remains 152 passed, 2 deselected.

- `scripts/generate_icon.py:36`: drop redundant `int(round(...))` —
  `round()` already returns `int` when called without a digits
  argument in Python 3 (RUF046).
- `src/soyle/app.py:170`: replace `try / except Exception / pass`
  around `keyboard.unhook()` with `contextlib.suppress(Exception)`.
  Same semantics, fewer lines, and the intent ("we don't care if
  the hook is already gone during shutdown") is now legible at the
  call site rather than implicit in the empty except block (SIM105).
- `tests/unit/test_shortcut_capture.py:2`: ruff --fix re-sorted the
  import block (I001). No content change, just ordering.

+++ scripts/generate_icon.py
@@ -33,7 +33,7 @@ def _render(size: int) -> Image.Image:
 
     # Everything below is authored in 256-unit virtual coords.
     def sc(v: float) -> int:
-        return int(round(v * size / 256))
+        return round(v * size / 256)
 
     # Background circle with a small transparent margin so the shape
     # doesn't crop on circular-badge OSes (macOS dock, some Windows skins).
